Compiler warnings w/ -m32 for coordinatorapi.cpp
 - Util::readAll() and Util::writeAll() return ssize_t, which is a
    signed int when compiled in 32-bit mode.  When the result is
    compared with a uint64_t, the gcc-4.8 compiler issues warnings
    about comparing a signed and unsigned int expression.  So, we
    cast the other comparison expression to an 'ssize_t'.

+++ src/coordinatorapi.cpp
@@ -867,7 +867,7 @@ int CoordinatorAPI::getUniqueIdFromCoordinator(const char *id,
   }
 
   JASSERT(Util::writeAll(sock, &msg, sizeof(msg)) == sizeof(msg));
-  JASSERT(Util::writeAll(sock, key, key_len) == key_len);
+  JASSERT(Util::writeAll(sock, key, key_len) == (ssize_t)key_len);
 
   msg.poison();
